+++ levior/ocresolvers.py
@@ -26,7 +26,10 @@ def random_wb_user_agent(os_list,
 OmegaConf.register_new_resolver("choice", choice)
 OmegaConf.register_new_resolver("random", choice)
 OmegaConf.register_new_resolver("random_user_agent", random_useragent)
+OmegaConf.register_new_resolver("ua_roulette", random_useragent)
 OmegaConf.register_new_resolver("rweb_user_agent",
+                                random_wb_user_agent)
+OmegaConf.register_new_resolver("custom_ua_roulette",
                                 random_wb_user_agent)
 OmegaConf.register_new_resolver(
     "datetime_now_iso",