+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add an OC resolver to generate a random user agent
 - Add an OC resolver to generate a custom random user agent with specific
   params (OS, software, software engine ...)
+- Configurable HTTP request headers
 
 ### Changed